asx_spider/requirements.txt
requests==2.19.1
beautifulsoup4==4.6.0
fake-useragent==0.1.10
mysql-connector==2.1.6
lxml==4.2.3
